requirements.txt
click>=8.1.0
requests>=2.31.0
openai>=1.12.0
pyyaml>=6.0.1
inquirer>=3.1.3
ctransformers>=0.2.27
python-dotenv>=1.0.0